Refactor transaction link creation
